requirements.txt
pandas
xlsxwriter
discord-py-interactions
